bump for my single question votes: random flow

+++ app/src/main/java/com/robandboo/fq/service/QuestionService.java
@@ -37,5 +37,5 @@ public interface QuestionService {
     Call<List<QuestionFile>> loadFile(@Path("id") Long id);
 
     @GET("/api/v1/question/get/file/{id}")
-    Call<List<QuestionFile>> getFileByQuestionId(@Path("id") int id);
+    Call<List<QuestionFile>> getFileByQuestionId(@Path("id") Long id);
 }
